Renames ExampleApp.Test.Functional to Slinqy.Test.Functional since this project is for testing Slinqy primarily, even though it uses the ExampleApp to do so.
