+++ Source/Slinqy.Test.Functional/Features/Infrastructure/Versioning.feature.cs
@@ -11,22 +11,22 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace ExampleApp.Test.Functional.Features
+namespace Slinqy.Test.Functional.Features.Infrastructure
 {
     using TechTalk.SpecFlow;
     
     
     [System.CodeDom.Compiler.GeneratedCodeAttribute("TechTalk.SpecFlow", "1.9.0.77")]
     [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    public partial class AutoExpandQueueStorageFeature : Xunit.IClassFixture<AutoExpandQueueStorageFeature.FixtureData>, System.IDisposable
+    public partial class VersionFeature : Xunit.IClassFixture<VersionFeature.FixtureData>, System.IDisposable
     {
         
         private static TechTalk.SpecFlow.ITestRunner testRunner;
         
-#line 1 "AutoExpandQueueStorage.feature"
+#line 1 "Versioning.feature"
 #line hidden
         
-        public AutoExpandQueueStorageFeature()
+        public VersionFeature()
         {
             this.TestInitialize();
         }
@@ -34,8 +34,8 @@ namespace ExampleApp.Test.Functional.Features
         public static void FeatureSetup()
         {
             testRunner = TechTalk.SpecFlow.TestRunnerManager.GetTestRunner();
-            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "AutoExpandQueueStorage", "As a Queue Writer\nI want my queue storage capacity to expand when it nears capaci" +
-                    "ty\nSo that I can always store my messages", ProgrammingLanguage.CSharp, ((string[])(null)));
+            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "Version", "As a User\r\nI want to know what version I\'m using\r\nSo that I can understand how th" +
+                    "e system should act", ProgrammingLanguage.CSharp, ((string[])(null)));
             testRunner.OnFeatureStart(featureInfo);
         }
         
@@ -64,7 +64,7 @@ namespace ExampleApp.Test.Functional.Features
             testRunner.CollectScenarioErrors();
         }
         
-        public virtual void SetFixture(AutoExpandQueueStorageFeature.FixtureData fixtureData)
+        public virtual void SetFixture(VersionFeature.FixtureData fixtureData)
         {
         }
         
@@ -74,19 +74,17 @@ namespace ExampleApp.Test.Functional.Features
         }
         
         [Xunit.FactAttribute()]
-        [Xunit.TraitAttribute("FeatureTitle", "AutoExpandQueueStorage")]
-        [Xunit.TraitAttribute("Description", "Queue Storage Expands When Utilization Reaches Threshold")]
-        public virtual void QueueStorageExpandsWhenUtilizationReachesThreshold()
+        [Xunit.TraitAttribute("FeatureTitle", "Version")]
+        [Xunit.TraitAttribute("Description", "Deployed Version Matches Test Version")]
+        public virtual void DeployedVersionMatchesTestVersion()
         {
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Queue Storage Expands When Utilization Reaches Threshold", ((string[])(null)));
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Deployed Version Matches Test Version", ((string[])(null)));
 #line 6
 this.ScenarioSetup(scenarioInfo);
 #line 7
- testRunner.Given("a Queue with Storage Utilization Scale Up Threshold set", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Given ");
+ testRunner.Given("I navigate to the Homepage", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Given ");
 #line 8
- testRunner.When("the Queue Storage Utilization reaches the Scale Up Threshold", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "When ");
-#line 9
- testRunner.Then("the Queue Storage Capacity expands", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Then ");
+ testRunner.Then("the Application Version matches the Test Version", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Then ");
 #line hidden
             this.ScenarioCleanup();
         }
@@ -98,12 +96,12 @@ this.ScenarioSetup(scenarioInfo);
             
             public FixtureData()
             {
-                AutoExpandQueueStorageFeature.FeatureSetup();
+                VersionFeature.FeatureSetup();
             }
             
             void System.IDisposable.Dispose()
             {
-                AutoExpandQueueStorageFeature.FeatureTearDown();
+                VersionFeature.FeatureTearDown();
             }
         }
     }
