+++ Source/Slinqy.Test.Functional/Models/ExampleAppPages/WebsiteFooter.cs
@@ -1,4 +1,4 @@
-﻿namespace ExampleApp.Test.Functional.Models.ExampleAppPages
+﻿namespace Slinqy.Test.Functional.Models.ExampleAppPages
 {
     using OpenQA.Selenium;
     using OpenQA.Selenium.Support.PageObjects;
