+++ Source/Slinqy.Test.Functional/Models/SeleniumWebBase.cs
@@ -1,4 +1,4 @@
-﻿namespace ExampleApp.Test.Functional.Models
+﻿namespace Slinqy.Test.Functional.Models
 {
     using OpenQA.Selenium;
     using OpenQA.Selenium.Support.PageObjects;
