+++ Source/Slinqy.Test.Functional/Models/WebBrowser.cs
@@ -1,7 +1,6 @@
-﻿namespace ExampleApp.Test.Functional.Models
+﻿namespace Slinqy.Test.Functional.Models
 {
     using OpenQA.Selenium;
-    using OpenQA.Selenium.Chrome;
     using System;
     using System.Collections.Generic;
     using System.Globalization;
