+++ Source/Slinqy.Test.Functional/Models/WebPage.cs
@@ -1,4 +1,4 @@
-﻿namespace ExampleApp.Test.Functional.Models
+﻿namespace Slinqy.Test.Functional.Models
 {
     using OpenQA.Selenium;
     using System;
