+++ Source/Slinqy.Test.Functional/Models/WebPageWidget.cs
@@ -1,4 +1,4 @@
-﻿namespace ExampleApp.Test.Functional.Models
+﻿namespace Slinqy.Test.Functional.Models
 {
     using OpenQA.Selenium;
