+++ Source/Slinqy.Test.Functional/Slinqy.Test.Functional.csproj
@@ -10,8 +10,8 @@
     <ProjectGuid>{5FB77DEA-B800-42CC-9A5F-673608CF9D12}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ExampleApp.Test.Functional</RootNamespace>
-    <AssemblyName>ExampleApp.Test.Functional</AssemblyName>
+    <RootNamespace>Slinqy.Test.Functional</RootNamespace>
+    <AssemblyName>Slinqy.Test.Functional</AssemblyName>
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
