+++ Source/Slinqy.Test.Functional/app.config
