+++ Source/Slinqy.Test.Functional/packages.config
