+++ Source/Slinqy.sln
@@ -25,12 +25,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{151D2E53-A2C4-4D7D-83FE-D05416EBD58E}") = "ExampleApp.ResourceGroup", "ExampleApp.ResourceGroup\ExampleApp.ResourceGroup.deployproj", "{80EC2E4E-7ACC-4219-BB95-AD80DB7DD318}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleApp.Test.Functional", "ExampleApp.Test.Functional\ExampleApp.Test.Functional.csproj", "{5FB77DEA-B800-42CC-9A5F-673608CF9D12}"
-	ProjectSection(ProjectDependencies) = postProject
-		{CF7FAD0F-EFD5-4FCE-8141-78A5CCD8FC77} = {CF7FAD0F-EFD5-4FCE-8141-78A5CCD8FC77}
-	EndProjectSection
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleApp.Web", "ExampleApp.Web\ExampleApp.Web.csproj", "{CF7FAD0F-EFD5-4FCE-8141-78A5CCD8FC77}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ExampleApp", "ExampleApp", "{EFED8FC3-CEF2-4BF7-AE02-8110D4AE0BA3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Slinqy.Test.Functional", "Slinqy.Test.Functional\Slinqy.Test.Functional.csproj", "{5FB77DEA-B800-42CC-9A5F-673608CF9D12}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,12 +38,16 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{80EC2E4E-7ACC-4219-BB95-AD80DB7DD318}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{80EC2E4E-7ACC-4219-BB95-AD80DB7DD318}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5FB77DEA-B800-42CC-9A5F-673608CF9D12}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5FB77DEA-B800-42CC-9A5F-673608CF9D12}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CF7FAD0F-EFD5-4FCE-8141-78A5CCD8FC77}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CF7FAD0F-EFD5-4FCE-8141-78A5CCD8FC77}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5FB77DEA-B800-42CC-9A5F-673608CF9D12}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5FB77DEA-B800-42CC-9A5F-673608CF9D12}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{80EC2E4E-7ACC-4219-BB95-AD80DB7DD318} = {EFED8FC3-CEF2-4BF7-AE02-8110D4AE0BA3}
+		{CF7FAD0F-EFD5-4FCE-8141-78A5CCD8FC77} = {EFED8FC3-CEF2-4BF7-AE02-8110D4AE0BA3}
 	EndGlobalSection
 EndGlobal
